Add MonsterLibrary to handle monster index and custom monsters
Moves the global monster list from MonsterTableModel to MonsterLibrary,
and adds ability to add custom monsters which are stored in a single xml
file. Custom monsters are created by editing the monster in the
encounter dialog and then saving to the library.
Changes stats block fields to be stored as properties with the "field"
prefix. This also fixes broken field editing in the encounter dialog due
to the DefaultDetailPanel not using the correct property name.