DTT: Enhance light sources
Option to always show the origin. Option to paint the lit area with a
specified colour. If set to radiate from all corners of token, ignore
the light source's position so that the light source appears locked to
the token. Also paint the origin indicator in all four corners of the
token.